requirements.txt
langchain
langchain-community
langchain-openai
openai
python-dotenv
wikipedia
google-search-results